Change how squasher works
Tween in script now updates the scale of the pillar and its hitbox, and the position of the top part of the pillar and hitbox. saves more space and easier to see in editor. might change it up again, but alright for now.
doesn't yield properly for now and bugs out after the first animation cycle but then it works after that

idk what the tileset thing is